@@ -1,6 +1,5 @@
 package com.in28minutes.springboot.learnspringboot;
 
-import java.util.Arrays;
 import java.util.List;
 
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,13 +10,12 @@ public class CourseController {
 	
 	@RequestMapping("/courses")
 	public List<Course> retrieveAllCourses() {
-		return Arrays.asList(
+		return List.of(
 				new Course(1, "Learn AWS", "in28minutes"),
 				new Course(2, "Learn DevOps", "in28minutes"),
 				new Course(3, "Learn Azure", "in28minutes"),
 				new Course(4, "Learn GCP", "in28minutes")
-				
-				);
+		);
 	}
 
 }
