@@ -35,7 +35,7 @@ public class Todo {
 
 	private String username;
 	
-	@Size(min=10, message="Enter atleast 10 characters")
+	@Size(min=10, message="Enter at least 10 characters")
 	private String description;
 	private LocalDate targetDate;
 	private boolean done;
