@@ -7,24 +7,20 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.SessionAttributes;
+import org.springframework.web.bind.annotation.*;
 
 import jakarta.validation.Valid;
 
 //@Controller
 @SessionAttributes("name")
 public class TodoController {
+
+	private TodoService todoService;
 	
 	public TodoController(TodoService todoService) {
 		super();
 		this.todoService = todoService;
 	}
-
-	private TodoService todoService;
-		
 	
 	@RequestMapping("list-todos")
 	public String listAllTodos(ModelMap model) {
@@ -87,8 +83,7 @@ public class TodoController {
 	}
 
 	private String getLoggedInUsername(ModelMap model) {
-		Authentication authentication = 
-				SecurityContextHolder.getContext().getAuthentication();
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 		return authentication.getName();
 	}
 
