@@ -5,5 +5,5 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TodoRepository extends JpaRepository<Todo, Integer>{
-	public List<Todo> findByUsername(String username);
+	List<Todo> findByUsername(String username);
 }
