@@ -12,7 +12,7 @@ import jakarta.validation.Valid;
 @Service
 public class TodoService {
 	
-	private static List<Todo> todos = new ArrayList<>();
+	private static final List<Todo> todos = new ArrayList<>();
 	
 	private static int todosCount = 0;
 	
