@@ -60,16 +60,15 @@ public class SpringSecurityConfiguration {
 	
 	@Bean
 	public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-		
-		
+
 		http.authorizeHttpRequests(
 				auth -> auth.anyRequest().authenticated());
 		
 		http.httpBasic(withDefaults());
 		
-		http.csrf().disable(); //POST or PUT
+		http.csrf(csrf -> csrf.disable()); //POST or PUT
 		
-		http.headers().frameOptions().disable();
+		http.headers(headers -> headers.frameOptions(frameOptionsConfig-> frameOptionsConfig.disable()));
 		
 		return http.build();
 	}
